fixes
fixed so we could use the whole trace-compile as the test, with using '-f', file as the command.
then also fixed the issue that we only showed 6 pages and I fixed the simulation process to get the 339 pages like we were supposed to.

now i have an interval of 50 but in the future before submission i will change to 25 or 10 perhaps, this will take around 12 to 30 minutes